RFID Implementation
After a lot of struggling, I added the possibility of also being able to
scan RFID cards. When an entered ID has a length of 16, it'll perform a
fallover method where it returns the username based on the RFID entered.
The username is found in the PCOUNTER database using the ACCOUNT.EXE
tool provided by PCOUNTER-EUROPE