Error display fixes
Better exception handling. Previously it would return "Could not
connect" to server if a user was not found, now it will properly display
this error if a user isn't found in the system. However it will also
display "User not found" if it is not inside the domain since the search
result will still be empty.
Also encased the screen protection function in try catch blocks to avoid
error throwing if a password is being reset while the screen is trying
to refresh.